rename /schema to /graphql
